+++ Editor/Attributes/IndentDrawer.cs
@@ -1,0 +1,22 @@
+using UnityEditor;
+using UnityEngine;
+
+namespace AnkleBreaker.Utils.Inspector.Editor
+{
+    [CustomPropertyDrawer(typeof(IndentAttribute))]
+    public class IndentDrawer : PropertyDrawer
+    {
+        public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
+        {
+            var attr = (IndentAttribute)attribute;
+            EditorGUI.indentLevel += attr.Level;
+            EditorGUI.PropertyField(position, property, label, true);
+            EditorGUI.indentLevel -= attr.Level;
+        }
+
+        public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
+        {
+            return EditorGUI.GetPropertyHeight(property, label, true);
+        }
+    }
+}