+++ Runtime/Attributes/TagFieldAttribute.cs
@@ -1,0 +1,11 @@
+using System;
+using UnityEngine;
+
+namespace AnkleBreaker.Utils.Inspector
+{
+    /// <summary>
+    /// Displays a string field as a Unity Tag dropdown.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Field, Inherited = true, AllowMultiple = false)]
+    public class TagFieldAttribute : PropertyAttribute { }
+}